refactor: GeocodingUtil을 GeocodingService로 리팩토링
