+++ src/main/java/edu/example/wayfarer/service/GeocodingService.java
@@ -1,0 +1,7 @@
+package edu.example.wayfarer.service;
+
+public interface GeocodingService {
+
+    String geocoding(String address);
+    String reverseGeocoding(double lat, double lng);
+}
